Developing the Karatsuba algorithm in Python
How long does it take to multiply two natural numbers? If the numbers have m and n digits respectively, the usual algorithm consumes time proportional to m × n.

1414213562373095048801688724209
×3141592653589793238462643383

It's difficult to imagine a faster algorithm than the usual one if m and n are small (a few hundred, say). But there is an algorithm that is much faster when m and n are large (as happens in cryptography, for example).

The Karatsuba algorithm (Карацуба) uses the divide-and-conquer method. Analysis of the algorithm's performance leads to an interesting recurrence.

This page teaches some general lessons: (1) for many problems, there are surprisingly faster algorithms than the usual algorithm; (2) sometimes the superiority of the more sophisticated algorithm only becomes apparent in very large instances of the problem; (3) recurrence resolution techniques are capable of performing a very precise analysis of the performance of recursive algorithms.